Functions

Oura.getdailyactivity — Function

getdailyactivity(client::OuraClient, startDate::String, endDate::String)

Get daily activity data for a given time period. default is to give the last day.

Oura.getdailycardiovascularage — Function

getdailycardiovascularage(client::OuraClient, startDate::String, endDate::String)

Get daily cardiovascular age data for a given time period. default is to give the last day.

Oura.getdailyreadiness — Function

getdailyreadiness(client::OuraClient, startDate::String, endDate::String)

Get daily readiness data for a given time period. default is to give the last day.

Oura.getheartrate — Function

getheartrate(client::OuraClient, startDate::String, endDate::String)

Get all heartrate data for a given time period. default is to give the last day.

Oura.getpersonalinfo — Method

getpersonalinfo(client::OuraClient)

Get all PersonalInfo

Oura.getsleep — Function

getsleep(client::OuraClient, startDate::String, endDate::String)

Get all sleep data for a given time period. default is to give the last day.

Oura.OuraClient — Type

OuraClient(token::String,url::String,sandbox::Bool)

Client used for identification. Simply contains the authentification token. Sandbox can be set to true to use fake data